chore consolidate and cleanup tx history code
