--- src/modules/history/downloadHelper.tsx
@@ -1,11 +1,0 @@
-export const downloadData = (fileName: string, content: string, mimeType: string) => {
-  const file = new Blob([content], { type: mimeType });
-  const downloadUrl = URL.createObjectURL(file);
-  const link = document.createElement('a');
-  link.href = downloadUrl;
-  link.download = fileName;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-  URL.revokeObjectURL(downloadUrl);
-};
